corrected bugs
